requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.0.0
PyMuPDF==1.23.7
python-dateutil==2.8.2
Flask-Session==0.5.0 